\name{gather}
\alias{gather}
\title{Utility function}
\description{
  Called by various functions to calculate 
  various intercepts.
}

\usage{
    gather(u, v, w, moran, df, xpoints, filter, fw)
}
\arguments{
  \item{u}{a vector.}
  \item{v}{a vector.}
 \item{w}{a vector.}
 \item{moran}{a matrix.}
\item{df}{a scalar.}
 \item{xpoints}{a vector.}
 \item{filter}{a logical.}
  \item{fw}{a scalar}
       }

\details{
An auxiliary function to ease  maintenance.
}

\value{
A listis returned.
} 
\keyword{misc}
