\name{gcdist}
\alias{gcdist}
\title{Greater circle distance}
\description{
  Greater circle distance function to calculate spatial distance from lat-long data.  Called by various functions.
}

\usage{
    gcdist(x, y)
}
\arguments{
  \item{x}{Vector of longitudes.}
  \item{y}{Vector of latitudes.}
}

\value{
The distance in km is returned
} 

\keyword{misc}
