% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCf.R
\name{generate_NCf}
\alias{generate_NCf}
\title{generate_NCf}
\usage{
generate_NCf(
  out_dir = "",
  environment_name = "NCf",
  overwrite = TRUE,
  chunksizes_list = c(time = 365),
  unlim_list = c("time"),
  verbose = FALSE
)
}
\arguments{
\item{out_dir}{*character, default=""* Name of the output directory for the NetCDF file.}

\item{environment_name}{*character, default="NCf"* Name of the R environment used to store NCf variables.}

\item{overwrite}{*logical, default=TRUE* If a NetCDF file already exists in the `out_dir` directory, will it be overwrite by the new generated one ?}

\item{chunksizes_list}{*named integer vector, default=c("time"=365)* Defines the `chunksizes` value of the [ncdf4::ncvar_def()] function for a defined variable explained as the name of this current vector value. For example, by default with `c("time"=365)` the dimension `time` will have a `chunksizes` of 365.}

\item{unlim_list}{*character vector, default=c("time")* Defines as `TRUE` the `unlim` variable of the [ncdf4::ncdim_def()] function for set of dimensions named listed in this `unlim_list` variables.}

\item{verbose}{*logical, default=FALSE* Prints basic info about the current execution in order to debug possible format issues.}
}
\value{
NetCDF file in the `out_dir` directory.
}
\description{
It creates a NetCDF file according to the defined variables in the environemment variable previously created with the [initialise_NCf()] function. Each variable name needs to follow a precise pattern in order to be understandable by this function.
}
\note{
# General Principle
---
The idea is to start with an empty environement variable created by [initialise_NCf()] and to store variable in this environnement variable. For example, the default environemment variable is named `NCf` so the variable `human` that take the character `"George"` can be add to `NCf` with `NCf$human = "George"`. In fact this variable will be un-used in the final NetCDF file because it does not indicate to which NetCDF variable or dimension the `human` info refers to. 

Thus, all  R variables that will be processed by the [generate_NCf()] function all have the same name format `NCf$xxxx.yyyy` or `NCf$xxxx.00.yyyy`. The latter format corresponds to a character string followed by a dollar, then an alphanumeric character string, followed by a period ".", then a two-digit number, followed by a new period "." and a new alphanumeric character string. That way, the first string `NCf` is the storage environment for variables specific to the current NetCDF build (... basically it's the NetCDF file ID) and it's the dollar to access it. The second string `xxxx` is the variable (or dimension) that is being discussed in the NetCDF and the third and last string `yyyy` is the attribute (or associated parameters) that we want to fill in. Finally, the number `00` is optional and is simply present to manage the order of appearance of the attributes in the final NetCDF.

Thus, adding a variable (or dimension) and modifying these attributes in the final NetCDF is simply a matter of modifying the name of an R variable and its associated value.

# Advanced principle
---
For a better understanding of how the code works it is important to specify some rules for editing R NCf variables.

## NetCDF dimension
Defining a NetCDF dimension is done by creating an R variable `NCf$x.name = "a"`. Here `a` is the name displayed in the NetCDF of the dimension and `x` is the internal R code identification of this variable. It is also necessary to define an R variable `NCf$x.value = Value` where this time `Value` is for example a data vector which defines the values associated with the dimension identified in R by `x` and named in the NetCDF `a`. For the sake of clarity, it is often best to define a dimension named in the NetCDF in the same way as its internal identification in R such as `NCf$y.name = "y"`.

## NetCDF variable
Defining a NetCDF variable is done in the same way as for a dimension via an R variable `NCf$var.name = "b"`. However, this time it is not necessary to associate another R variable of type `NCf$var.value = Value` but an R variable of type `NCf$var.dimension = "x"`. This last one allows to make the link between the variable `b` and the dimension `x`. If the dimension entered is "", no dimension will be associated with this variable but this declaration in R is still necessary.

## Variable precision
For a variable (and more rarely for a dimension), it is possible to give the "precision" or rather the type of data associated with it. To do this, you have to define an R variable `NCf$x.precision = "type"` where `"type"` is chosen among 'short', 'integer', 'float', 'double', 'char' and 'byte'. WARNING ... if a character variable (thus of type 'char') is filled in and takes as input a dimension characterizing the length of this character string, it is imperative that the associated dimension takes as value a vector of 1 to the desired length of the character string and that this same dimension presents a parameter `NCf$dim.is_nchar_dimension = TRUE` which specifies this special behavior.

## Variable attribute
Any other attribute of a NetCDF variable or dimension is defined by an R variable `NCf$var.00.att = "attribute"` where in general `"attribute"` is a string and preferably a two-digit number (in this case `00`) is used to specify its position in the NetCDF.

## Global attribute
A NetCDF attribute defined as `NCf$global.00.att = "attribute"` specifies with the `global` tag that this attribute is global in the NetCDF.

## NetCDF file title
A NetCDF attribute defined as `NCf$title.00.att = "attribute"` specifies with the `title` marker that this attribute allows the construction of the NetCDF file title by joining with `"_"` the set of non-empty supplied attributes.

# Framework
---
In that way, NetCDF file creation is more focused on the info and formatting than on the raw code of NetCDF creation. It is easy to manage multiple NetCDF file at the same time with different environment variables and to have separate scripts that contain NCf R variables definitions between R code line of data processing.
}
\examples{
## 1. INITIALISATION _________________________________________________
initialise_NCf()


## 2. TITLE __________________________________________________________
NCf$title.01.title = paste0("MODEL_", Sys.Date())


## 3. GLOBAL ATTRIBUTS _______________________________________________
NCf$global.01.data_type = "diagnostic"
NCf$global.02.contact = "@"


## 4. DIMENSIONS _____________________________________________________
### 4.1. Time ________________________________________________________
start = "2000-01-01"
end = "2000-01-31"
timezone = "UTC"
step = "days"

from = as.POSIXct(start, tz=timezone)
to = as.POSIXct(end, tz=timezone)
origin = as.POSIXct("1950-01-01", tz=timezone)
units = paste0(step, " since ", origin)
time = seq.POSIXt(from=from, to=to, by=step)
time = as.numeric(time - origin)

NCf$time.name = "time"
NCf$time.value = time
NCf$time.01.standard_name = "time"
NCf$time.02.units = units

### 4.2. Station _____________________________________________________
NCf$station.name = "station"
NCf$station.value = 1:3

NCf$code.name = "code"
NCf$code.dimension = "station, code_strlen"
NCf$code.precision = "char"
NCf$code.value = c("AAAAAAAA", "BBBBBBBB", "CCCCCCCC")
NCf$code.01.standard_name = "code"
NCf$code_strlen.name = "code_strlen"
NCf$code_strlen.value = 1:max(nchar(NCf$code.value))
NCf$code_strlen.is_nchar_dimension = TRUE


## 5. VARIABLES ______________________________________________________
### 5.1. Flow ________________________________________________________
NCf$Q.name = "Q"
NCf$Q.dimension = "station, time"
NCf$Q.precision = "float"
NCf$Q.value = matrix(
    data=round(x=runif(length(NCf$time.value)*length(NCf$station.value)),
               digits=2),
    ncol=length(NCf$time.value))
NCf$Q.01.standard_name = "flow"
NCf$Q.02.units = "m3.s-1"
NCf$Q.03.missing_value = NaN

NCf$surface_model.name = "surface_model"
NCf$surface_model.dimension = "station"
NCf$surface_model.precision = "float"
NCf$surface_model.value = round(x=runif(length(NCf$station.value)),
                                digits=2)
NCf$surface_model.01.standard_name = "surface in the model world"
NCf$surface_model.02.units = "km2"
NCf$surface_model.03.missing_value = NaN


## 6. SAVING _________________________________________________________
generate_NCf(out_dir="./")
}
\seealso{
[GitHub developpement page](https://github.com/super-lou/NCf) and [an advanced use of the NCf package](https://github.com/super-lou/Ex2D_toolbox/tree/main/help/DRIAS_export/DRIAS_export_1D)
}
