% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbm}
\alias{lbm}
\title{Spatio-temporal data panel of Larch Budmoth defoliation}
\format{A data-frame with 135 rows and 40 columns. The first two are the x- and y-coordinates (in m), the following 38 represents the defoliation in years 1961 through 1998.}
\usage{
data(lbm)
}
\description{
This is the data in Bjornstad et al. (2002).
}
\references{
Bjornstad, O.N., Peltonen, M., Liebhold, A.M., and Baltensweiler, W. (2002) Waves of larch budmoth outbreaks in the European Alps. Science, 298, 1020-1023. \url{https://doi.org/10.1126/science.1075182}
}
\keyword{datasets}
