\name{print.Sncf}
\alias{print.Sncf}
\title{Print function for Sncf objects}
\description{
  `Print' method for class "Sncf".  
}

\usage{
    \method{print}{Sncf}(x, \dots)
}
\arguments{
  \item{x}{an object of class "Sncf", usually, as
  a result of a call to Sncf or related).}
  \item{\dots}{other arguments}
}

\value{
The function-call is printed to screen.
} 

\seealso{
\code{\link{Sncf}} 
}

