\name{print.Sncf2D}
\alias{print.Sncf2D}
\title{Print function for Sncf objects}
\description{
  `Print' method for class "Sncf2D".  
}

\usage{
    \method{print}{Sncf2D}(x, \dots)
}
\arguments{
  \item{x}{an object of class "Sncf2D", usually, as
  a result of a call to Sncf2D or spline.correlog.2D).}
  \item{\dots}{other arguments}
}

\value{
The function-call is printed to screen.
} 

\seealso{
\code{\link{Sncf}} 
}

