\name{print.spline.correlog}
\alias{print.spline.correlog}
\title{Print function for spline.correlog objects}
\description{
  `Print' method for class "spline.correlog".  
}

\usage{
    \method{print}{spline.correlog}(x, \dots)
}
\arguments{
  \item{x}{an object of class "spline.correlog", usually, as
  a result of a call to spline.correlog or related).}
  \item{\dots}{other arguments}
}

\value{
The function-call is printed to screen.
} 

\seealso{
\code{\link{spline.correlog}} 
}

