% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sncf.R
\name{summary.Sncf}
\alias{summary.Sncf}
\title{Summarizing nonparametric spatial correlation-functions}
\usage{
\method{summary}{Sncf}(object, ...)
}
\arguments{
\item{object}{an object of class "Sncf", usually, as a result of a call to \code{\link{Sncf}} (or \code{\link{Sncf.srf}}).}

\item{\dots}{other arguments}
}
\value{
A list summarizing the nonparametric (cross-)covariance function is returned. 
\item{Regional.synch}{the regional mean (cross-)correlation.}
\item{Squantile}{the quantile distribution from the resampling for the regional correlation.}
\item{estimates}{a vector of benchmark statistics:}
\item{$x}{is the lowest value at which the function is = 0. If correlation is initially negative, the distance calculated appears as a negative measure.}
\item{$e}{is the lowest value at which the function is <= 1/e.}
\item{$y}{is the extrapolated value at x=0.} 
\item{$cbar}{is the shortest distance at which function is = regional mean correlation.}
\item{quantiles}{a matrix summarizing the quantiles in the bootstrap (or null) distributions of the benchmark statistics.}
}
\description{
'summary' method for class "Sncf".
}
\seealso{
\code{\link{Sncf}}, \code{\link{plot.Sncf}}
}
\keyword{regression}
\keyword{smooth}
