% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ff.filter}
\alias{ff.filter}
\title{Fourier filter for c'fns}
\usage{
ff.filter(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A vector is returned whose Fourier-transform has no non-negative coeficients.
}
\description{
Fourier filter to ensure positive semi-definite correlation fnctions.  Called by various functions.
}
\seealso{
\code{\link{Sncf}}
}
\keyword{misc}
