% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gather}
\alias{gather}
\title{Utility function}
\usage{
gather(u, v, w, moran, df, xpoints, filter, fw)
}
\arguments{
\item{u}{a vector.}

\item{v}{a vector.}

\item{w}{a vector.}

\item{moran}{a matrix.}

\item{df}{a scalar.}

\item{xpoints}{a vector.}

\item{filter}{a logical.}

\item{fw}{a scalar}
}
\value{
A list is returned.
}
\description{
Called by various functions to calculate various intercepts.
}
\details{
An auxiliary function to ease  maintenance.
}
\keyword{misc}
