% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gcdist}
\alias{gcdist}
\title{Great-circle distance}
\usage{
gcdist(x, y)
}
\arguments{
\item{x}{vector of longitudes.}

\item{y}{vector of latitudes.}
}
\value{
The distance in km is returned
}
\description{
Great-circle distance function to calculate spatial distance from lat-long data.  Called by various functions.
}
\keyword{misc}
