% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCf.R
\name{generate_NCf}
\alias{generate_NCf}
\title{generate_NCf}
\usage{
generate_NCf(
  out_dir = "",
  environment_name = "NCf",
  overwrite = TRUE,
  chunksizes_list = c(time = 365),
  verbose = FALSE
)
}
\arguments{
\item{out_dir}{Name of the output directory for the NetCDF file}

\item{environment_name}{Name of the R environment used to store
NCf variables}
}
\value{
NetCDF file in the 'out_dir' directory
}
\description{
Initialises the NCf way of NetCDF creation
}
\examples{
generate_NCf()
}
