% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCf.R
\name{initialise_NCf}
\alias{initialise_NCf}
\title{initialise_NCf}
\usage{
initialise_NCf(environment_name = "NCf")
}
\arguments{
\item{environment_name}{Name of the R environment used to store
NCf variables}
}
\value{
Nothing
}
\description{
Initialises the NCf way of NetCDF creation
}
\examples{
initialise_NCf()
}
