% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sncf.R
\name{plot.Sncf.cov}
\alias{plot.Sncf.cov}
\title{Plots nonparametric spatial covariance-functions}
\usage{
\method{plot}{Sncf.cov}(x, ...)
}
\arguments{
\item{x}{an object of class "Sncf.cov", usually, as a result of a call to \code{Sncf.srf} (with \code{corr} = FALSE).}

\item{\dots}{other arguments}
}
\value{
A plot of the nonparametric spatial covariance function (with CI's if boostrapps are available)
}
\description{
`plot' method for class "Sncf.cov".
}
\seealso{
\code{\link{Sncf.srf}}, \code{\link{plot.Sncf}}
}
\keyword{regression}
\keyword{smooth}
