% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sncf.R
\name{plot.Sncf.cov}
\alias{plot.Sncf.cov}
\title{Plots nonparametric spatial covariance-functions}
\usage{
\method{plot}{Sncf.cov}(x, xmax = 0, ...)
}
\arguments{
\item{x}{an object of class "Sncf.cov", usually, as a result of a call to \code{Sncf.srf} (with \code{corr} = FALSE).}

\item{xmax}{the maximal distance to be plotted on the x-axis. If set to zero the maximum distance in the data will be used.}

\item{\dots}{other arguments}
}
\description{
`plot' method for class "Sncf.cov".
}
\seealso{
\code{\link{Sncf.srf}}, \code{\link{plot.Sncf}}
}
\keyword{regression}
\keyword{smooth}
