% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sncf.R
\name{plot.Sncf}
\alias{plot.Sncf}
\title{Plots nonparametric spatial correlation-functions}
\usage{
\method{plot}{Sncf}(x, ylim = c(-1, 1), add = FALSE, ...)
}
\arguments{
\item{x}{an object of class "Sncf", usually, as a result of a call to \code{Sncf} (or \code{Sncf.srf}).}

\item{ylim}{limits for the y-axis (default: -1, 1).}

\item{add}{If TRUE the plot is added on to the previous graph.}

\item{\dots}{other arguments}
}
\value{
A plot of the nonparametric spatial covariance function (with CI's if bootstraps are available)
}
\description{
'plot' method for class "Sncf".
}
\seealso{
\code{\link{Sncf}}, \code{\link{Sncf.srf}}
}
\keyword{regression}
\keyword{smooth}
