% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-correlog.R
\name{plot.spline.correlog}
\alias{plot.spline.correlog}
\title{Plots spline correlograms}
\usage{
\method{plot}{spline.correlog}(x, xmax = 0, ylim = c(-1, 1), ...)
}
\arguments{
\item{x}{an object of class "spline.correlog", usually, as a result of a call to \code{spline.correlog}.}

\item{xmax}{the maximal distance to be plotted on the x-axis. If set to zero the maximum distance in the data will be used.}

\item{ylim}{limits for the y-axis (default: -1, 1).}

\item{\dots}{other arguments}
}
\value{
A plot of the spline correlogram function against distance is produced. 95\% pointwise confidence (or null) envelopes are superimposed (if available).
}
\description{
`plot' method for class "spline.correlog".
}
\seealso{
\code{\link{spline.correlog}}, \code{\link{summary.spline.correlog}}
}
\keyword{regression}
\keyword{smooth}
