% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSynch.R
\name{print.mSynch}
\alias{print.mSynch}
\title{Print function for mSynch objects}
\usage{
\method{print}{mSynch}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{an object of class "mSynch", usually, as a result of a call to \code{mSynch}.}

\item{verbose}{If TRUE, a raw listing of the object is produced. If FALSE, a summary list is produced}

\item{\dots}{other arguments}
}
\value{
If verbose is FALSE, a list summarizing the regional correlation is produced: 
\item{mean}{the regional mean correlation.} 
\item{Squantile}{the quantile distribution from the resampling for the regional correlation.}
}
\description{
`print' method for class "mSynch".
}
\seealso{
\code{\link{mSynch}}
}
\keyword{export}
\keyword{regression}
\keyword{smooth}
