% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sncf.R
\name{print.Sncf}
\alias{print.Sncf}
\title{Print function for Sncf objects}
\usage{
\method{print}{Sncf}(x, ...)
}
\arguments{
\item{x}{an object of class "Sncf", usually, as a result of a call to \code{Sncf} or related).}

\item{\dots}{other arguments}
}
\value{
The function-call is printed to screen.
}
\description{
'print' method for class "Sncf".
}
\seealso{
\code{\link{Sncf}}
}
