% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sncf2D.R
\name{print.Sncf2D}
\alias{print.Sncf2D}
\title{Print function for Sncf2D objects}
\usage{
\method{print}{Sncf2D}(x, ...)
}
\arguments{
\item{x}{an object of class "Sncf2D", usually, as a result of a call to \code{Sncf2D} or \code{spline.correlog2D}).}

\item{\dots}{other arguments}
}
\value{
The function-call is printed to screen.
}
\description{
`print' method for class "Sncf2D".
}
\seealso{
\code{\link{Sncf2D}}
}
