% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-correlog.r
\name{print.spline.correlog}
\alias{print.spline.correlog}
\title{Print function for spline.correlog objects}
\usage{
\method{print}{spline.correlog}(x, ...)
}
\arguments{
\item{x}{an object of class "spline.correlog", usually, as a result of a call to \code{spline.correlog} or related).}

\item{\dots}{other arguments}
}
\value{
The function-call is printed to screen.
}
\description{
`print' method for class "spline.correlog".
}
\seealso{
\code{\link{spline.correlog}}
}
