% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sncf2D.R
\name{summary.Sncf2D}
\alias{summary.Sncf2D}
\title{Summarizing anisotropic spatial correlation-functions}
\usage{
\method{summary}{Sncf2D}(object, ...)
}
\arguments{
\item{object}{an object of class "Sncf2D", usually, as a result of a call to \code{\link{Sncf2D}}.}

\item{\dots}{other arguments}
}
\value{
A list summarizing the nonparametric covariance function in each cardinal direction results, each with the entires as in \code{\link{summary.Sncf}}.
}
\description{
`summary' method for class "Sncf2D".
}
\seealso{
\code{\link{Sncf2D}}, \code{\link{cc.offset}}, \code{\link{summary.Sncf}}
}
\keyword{regression}
\keyword{smooth}
