% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-correlog.r
\name{summary.spline.correlog}
\alias{summary.spline.correlog}
\title{Summarizing spline correlograms}
\usage{
\method{summary}{spline.correlog}(object, ...)
}
\arguments{
\item{object}{an object of class "spline.correlog", usually, as a result of a call to \code{\link{spline.correlog}}.}

\item{\dots}{other arguments}
}
\value{
A list summarizing the spline correlogram is returned. 
\item{estimates}{a vector of benchmark statistics:}
\item{$x}{is the lowest value at which the function is = 0. If correlation is initially negative, the distance calculated appears as a negative measure.}
\item{$e}{is the lowest value at which the function is <= 1/e.}
\item{$y}{is the extrapolated value at x=0.} 
\item{quantiles}{a matrix summarizing the quantiles in the bootstrap (or null) distributions of the benchmark statistics.}
}
\description{
`summary' method for class "spline.correlog".
}
\seealso{
\code{\link{spline.correlog}}, \code{\link{plot.spline.correlog}}
}
\keyword{regression}
\keyword{smooth}
